tests: fix integration tests and rename binary
