+++ .dockerignore
@@ -3,3 +3,4 @@ test/e2e/build
 test/e2e/networks
 test/logs
 test/p2p/data
+third_party/bls-signatures/build
