+++ DOCKER/Makefile
@@ -5,9 +5,9 @@ push:
 	@sh -c "'$(CURDIR)/push.sh'"
 
 build_testing:
-	docker build --tag dashpay/tenderdash:testing -f ./Dockerfile.testing .
+	docker build --tag dashpay/tenderdash:testing -f ./Dockerfile.testing ..
 
 build_amazonlinux_buildimage:
-	docker build -t "dashpay/tenderdash:build_c-amazonlinux" -f Dockerfile.build_c-amazonlinux .
+	docker build -t "dashpay/tenderdash:build_c-amazonlinux" -f Dockerfile.build_c-amazonlinux ..
 
 .PHONY: build push build_testing build_amazonlinux_buildimage
