+++ DOCKER/build.sh
@@ -16,5 +16,5 @@ read -p "==> Build 3 docker images with the following tags (latest, $TAG, $TAG_N
 echo
 if [[ $REPLY =~ ^[Yy]$ ]]
 then
-		docker build -t "dashpay/tenderdash" -t "dashpay/tenderdash:$TAG" -t "dashpay/tenderdash:$TAG_NO_PATCH" .
+		docker build -t "dashpay/tenderdash" -t "dashpay/tenderdash:$TAG" -t "dashpay/tenderdash:$TAG_NO_PATCH" ..
 fi
