+++ cmd/tenderdash/commands/debug/debug.go
