+++ cmd/tenderdash/commands/debug/kill.go
