+++ cmd/tenderdash/commands/debug/util.go
