+++ cmd/tenderdash/commands/gen_validator.go
