+++ cmd/tenderdash/commands/init.go
