+++ cmd/tenderdash/commands/reset_priv_validator.go
