+++ cmd/tenderdash/commands/root_test.go
