+++ cmd/tenderdash/commands/show_node_id.go
