+++ cmd/tenderdash/commands/show_validator.go
